Read binary file using pkgutil.get_data
When running as a [zipapp][1] it's not possible to read package data
directly from the file system, as the zip file is not extracted. Using
[`pkgutil.get_data`][2] works, making the library compatible with being
installed in a zipapp.

One downside with this is that the entire file is ready into memory
instead of being parsed while on disk, but as the file is only 1.4 MB
this probably shouldn't be a big issue.

[1]: https://docs.python.org/3/library/zipapp.html
[2]: https://docs.python.org/3/library/pkgutil.html#pkgutil.get_data