Fix linting build failure and update supported Py versions
For _Black_ to run under _Python 3.14_, which is what the
[workflows/cron.yml][1] `python-version: 3.x` currently translates
into, we need to update _Black_ to its most recent version, which in
turn requires at least _Python 3.9_.

Yes, it ought to be possible to separate our the package's
requirements from the linting requirements. Still, feels both safest
and most straightforward to keep them the same. Especially given that
all the dropped Python versions have been unsupported for at least a
year.

Ought to resolve #16.

[1]: https://github.com/sethmlarson/hstspreload/blob/main/.github/workflows/cron.yml